+++ code/packages/rust/closure-pass-fold-control-flow/src/lib.rs
@@ -514,7 +514,8 @@ fn fold_expression(expr: &Expression, st: &mut FoldState) -> Expression {
         | Expression::StringLiteral(_)
         | Expression::BooleanLiteral(_)
         | Expression::NullLiteral(_)
-        | Expression::BigIntLiteral(_) => expr.clone(),
+        | Expression::BigIntLiteral(_)
+        | Expression::UndefinedLiteral(_) => expr.clone(),
 
         Expression::BinaryExpression(b) => Expression::BinaryExpression(BinaryExpression {
             cv: b.cv.clone(),